+++ qgreenland/tasks/common/raster.py
@@ -7,7 +7,10 @@ import rasterio as rio
 from qgreenland.constants import CONFIG, TaskType
 from qgreenland.util.luigi import LayerTask
 from qgreenland.util.misc import find_single_file_by_ext, temporary_path_dir
-from qgreenland.util.raster import gdal_calc_raster, warp_raster
+from qgreenland.util.raster import (gdal_calc_raster,
+                                    gdal_edit_raster,
+                                    gdal_mdim_translate_raster,
+                                    warp_raster)
 
 
 class BuildRasterOverviews(LayerTask):
@@ -68,6 +71,12 @@ class WarpRaster(LayerTask):
 
         extent_str = self.layer_cfg.get('extent', 'background')
         extent = CONFIG['project']['extents'][extent_str]
+
+        # TODO: we should probably do an intersection between the dataset bounds
+        # and the extent bounds to determine the correct values to pass to
+        # 'outputBounds'. If the dataset is SMALLER than the outputBounds, gdal
+        # will create a BIGGER dataset than the input, and fill it with e.g.,
+        # zeros
         warp_kwargs = {
             'resampleAlg': 'bilinear',
             'outputBounds': list(extent.values()),
@@ -89,7 +98,6 @@ class WarpRaster(LayerTask):
 class GdalCalcRaster(LayerTask):
 
     task_type = TaskType.WIP
-    input_ext_override = luigi.Parameter(default=None)
 
     def output(self):
         return luigi.LocalTarget(os.path.join(self.outdir, 'calc'))
@@ -100,6 +108,53 @@ class GdalCalcRaster(LayerTask):
             inp_path = find_single_file_by_ext(self.input().path,
                                                ext=self.layer_cfg['file_type'])
             gdal_calc_kwargs = self.layer_cfg['gdal_calc_kwargs']
-            gdal_calc_raster(inp_path, out_path,
-                             layer_cfg=self.layer_cfg,
-                             gdal_calc_kwargs=gdal_calc_kwargs)
+            gdal_calc_raster(
+                inp_path, out_path,
+                layer_cfg=self.layer_cfg,
+                gdal_calc_kwargs=gdal_calc_kwargs
+            )
+
+
+class GdalMDimTranslate(LayerTask):
+
+    task_type = TaskType.WIP
+    input_ext_override = luigi.Parameter(default=None)
+
+    def output(self):
+        return luigi.LocalTarget(os.path.join(self.outdir, 'gdal_mdim_translate'))
+
+    def run(self):
+        with temporary_path_dir(self.output()) as tmp_dir:
+            out_path = os.path.join(tmp_dir, self.filename)
+
+            file_ext = self.input_ext_override or self.layer_cfg['file_type']
+            inp_path = find_single_file_by_ext(self.input().path, ext=file_ext)
+
+            gdal_mdim_translate_kwargs = self.layer_cfg['gdal_mdim_translate_kwargs']
+            gdal_mdim_translate_raster(
+                inp_path, out_path,
+                layer_cfg=self.layer_cfg,
+                gdal_mdim_translate_kwargs=gdal_mdim_translate_kwargs
+            )
+
+
+class GdalEdit(LayerTask):
+
+    task_type = TaskType.WIP
+
+    def output(self):
+        return luigi.LocalTarget(os.path.join(self.outdir, 'gdal_edit'))
+
+    def run(self):
+        with temporary_path_dir(self.output()) as tmp_dir:
+            out_path = os.path.join(tmp_dir, self.filename)
+            inp_path = find_single_file_by_ext(self.input().path,
+                                               ext=self.layer_cfg['file_type'])
+            shutil.copy(inp_path, out_path)
+
+            gdal_edit_kwargs = self.layer_cfg['gdal_edit_kwargs']
+            gdal_edit_raster(
+                out_path,
+                layer_cfg=self.layer_cfg,
+                gdal_edit_kwargs=gdal_edit_kwargs
+            )