+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fix: Remove extraneous `valid.shp` shapefiles from final package.
 - Fix: "geothermal heat flux" layer units (MW/m2 -> mW/m2)
+- New "Glaciology" layer: Likely Basal Thermal State
 
 # v0.35.0 (2020-09-02)